requirements.txt
Flask==3.1.1
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pillow==11.2.1
python-dotenv==1.1.1
requests==2.32.4
setuptools==80.9.0
six==1.17.0
urllib3==2.5.0
Werkzeug==3.1.3
gunicorn